% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{attribute_count}
\alias{attribute_count}
\title{Get number of attributes for an item}
\usage{
attribute_count(delta, itemnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}
}
\value{
Integer number of attributes
}
\description{
Get number of attributes for an item
}
