% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{attribute_string}
\alias{attribute_string}
\title{Get attribute string}
\usage{
attribute_string(delta, itemnum, attrnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}

\item{attrnum}{Attribute number}
}
\value{
Character string with the attribute
}
\description{
Get attribute string
}
