% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{char_count}
\alias{char_count}
\title{Get number of characters}
\usage{
char_count(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Integer number of characters
}
\description{
Get number of characters
}
