% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{char_feature}
\alias{char_feature}
\title{Get character feature}
\usage{
char_feature(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Character string with the feature description
}
\description{
Get character feature
}
