% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{char_type}
\alias{char_type}
\title{Get character type}
\usage{
char_type(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Integer character type code
}
\description{
Get character type
}
