% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{char_type_name}
\alias{char_type_name}
\title{Get character type name}
\usage{
char_type_name(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Character string with type name
}
\description{
Get character type name
}
