% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{characters_df}
\alias{characters_df}
\title{Get character information as a data frame}
\usage{
characters_df(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
A data.frame with character information
}
\description{
Extracts all character definitions from a DELTA dataset and returns them as a
data frame.
}
