% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{chars_filename}
\alias{chars_filename}
\title{Get characters filename}
\usage{
chars_filename(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with the filename
}
\description{
Get characters filename
}
