% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkc.R
\name{checkc}
\alias{checkc}
\title{Check the characters}
\usage{
checkc(delta, remove_temp = TRUE)
}
\arguments{
\item{delta}{A Delta dataset object (created by \code{\link{load_delta}})}

\item{remove_temp}{Logical; if TRUE, remove temporary files after conversion (default = TRUE)}
}
\value{
Integer: 0 if check is successful, 1 if check fails
}
\description{
This function creates a directives file for the CONFOR program to check
the characters in a DELTA dataset.
}
\details{
The function assumes that the CONFOR program executable is available
in the system PATH or in the current working directory.
}
\examples{
\dontrun{
# Load a DELTA dataset
delta <- load_delta("chars_viola", "items_viola", "specs_viola")

# Check characters only
checkc(delta)
}
}
