% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checki.R
\name{checki}
\alias{checki}
\title{Check the items}
\usage{
checki(delta, remove_temp = TRUE)
}
\arguments{
\item{delta}{A Delta dataset object (created by \code{\link{load_delta}})}

\item{remove_temp}{Logical; if TRUE, remove temporary files after conversion (default = TRUE)}
}
\value{
Integer: 0 if check is successful, 1 if check fails
}
\description{
This function creates a directives file for the CONFOR program to check
the items in a DELTA dataset.
}
\details{
The function assumes that the CONFOR program executable is available
in the system PATH or in the current working directory.
}
\examples{
\dontrun{
# Load a DELTA dataset
delta <- load_delta("chars_viola", "items_viola", "specs_viola")

# Check items only
checki(delta)
}
}
