% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta_char_types}
\alias{delta_char_types}
\title{Get character type constants}
\usage{
delta_char_types()
}
\value{
Named integer vector of DELTA character types
}
\description{
Get character type constants
}
