% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{dependent_characters}
\alias{dependent_characters}
\title{Get all dependent characters}
\usage{
dependent_characters(delta, ccnum, ccstate)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}
}
\value{
Integer vector of dependent character numbers
}
\description{
Get all dependent characters
}
