% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{dependent_count}
\alias{dependent_count}
\title{Get number of dependent characters}
\usage{
dependent_count(delta, ccnum, ccstate)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}
}
\value{
Integer count of dependent characters
}
\description{
Get number of dependent characters
}
