% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{find_matching}
\alias{find_matching}
\title{Find all matching items}
\usage{
find_matching(delta, charnum, values, strict = TRUE, with_extrval = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{values}{Numeric vector of values}

\item{strict}{Logical; strict comparison}

\item{with_extrval}{Logical; include extreme values}
}
\value{
Integer vector of matching item numbers
}
\description{
Find all matching items
}
