% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{first_matching}
\alias{first_matching}
\title{Find first matching item}
\usage{
first_matching(delta, charnum, values, strict = TRUE, with_extrval = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{values}{Numeric vector of values}

\item{strict}{Logical; strict comparison}

\item{with_extrval}{Logical; include extreme values}
}
\value{
Integer item number or 0 if none
}
\description{
Find first matching item
}
