% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_methods.R
\name{get_all_depchar}
\alias{get_all_depchar}
\title{Get all dependent characters}
\usage{
get_all_depchar(delta, ccnum, ccstate)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}
}
\value{
Integer vector of dependent character numbers
}
\description{
Get all dependent characters
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_all_depchar(delta, 13, 2)
}
}
