% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_methods.R
\name{get_all_dependencies}
\alias{get_all_dependencies}
\title{Get all dependency relationships}
\usage{
get_all_dependencies(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
List of all dependency relationships
}
\description{
Get all dependency relationships
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_all_dependencies(delta)
}
}
