% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_methods.R
\name{get_all_implicit_values}
\alias{get_all_implicit_values}
\title{Get all implicit values}
\usage{
get_all_implicit_values(delta, iv_type = 1)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{iv_type}{Implicit value type (1 or 2, default: 1)}
}
\value{
Integer vector of implicit values
}
\description{
Get all implicit values
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_all_implicit_values(delta)
}
}
