% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_attribute}
\alias{get_attribute}
\title{Get attribute string (C++ style)}
\usage{
get_attribute(delta, itemnum, attrnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}

\item{attrnum}{Attribute number}
}
\value{
Character string with the attribute
}
\description{
Get attribute string (C++ style)
}
