% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_attributes_nb}
\alias{get_attributes_nb}
\title{Get number of attributes (C++ style)}
\usage{
get_attributes_nb(delta, itemnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}
}
\value{
Integer number of attributes
}
\description{
Get number of attributes (C++ style)
}
