% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_char_feature}
\alias{get_char_feature}
\title{Get character feature (C++ style)}
\usage{
get_char_feature(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Character string with the feature description
}
\description{
Get character feature (C++ style)
}
