% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_char_unit}
\alias{get_char_unit}
\title{Get character unit (C++ style)}
\usage{
get_char_unit(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Character string with the unit
}
\description{
Get character unit (C++ style)
}
