% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_chars_debug}
\alias{get_chars_debug}
\title{Get character debug output}
\usage{
get_chars_debug(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with debug output
}
\description{
Get character debug output
}
