% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_chars_nb}
\alias{get_chars_nb}
\title{Get number of characters (C++ style)}
\usage{
get_chars_nb(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Integer number of characters
}
\description{
Get number of characters (C++ style)
}
