% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_methods.R
\name{get_depchar}
\alias{get_depchar}
\title{Get dependent character by rank (C++ style)}
\usage{
get_depchar(delta, ccnum, ccstate, rank = 1)

get_depchar(delta, ccnum, ccstate, rank = 1)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}

\item{rank}{Rank position (default: 1)}
}
\value{
Integer dependent character number

Dependent character number or 0 if none
}
\description{
Get dependent character by rank (C++ style)

Get dependent character at specific rank
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_depchar(delta, 13, 2, 1)
}
}
