% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_methods.R
\name{get_depchar_nb}
\alias{get_depchar_nb}
\title{Get number of dependent characters (C++ style)}
\usage{
get_depchar_nb(delta, ccnum, ccstate)

get_depchar_nb(delta, ccnum, ccstate)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}
}
\value{
Integer count of dependent characters

Number of dependent characters
}
\description{
Get number of dependent characters (C++ style)

Get number of dependent characters
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_depchar_nb(delta, 13, 2)
}
}
