% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_example_files}
\alias{get_example_files}
\title{Get example file paths for testing}
\usage{
get_example_files(dataset = "viola")
}
\arguments{
\item{dataset}{Character string: "viola" or "grass"}
}
\value{
List with file paths
}
\description{
Get example file paths for testing
}
