% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_filename}
\alias{get_filename}
\title{Get filename (C++ style)}
\usage{
get_filename(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with the filename
}
\description{
Get filename (C++ style)
}
