% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_methods.R
\name{get_implicit_value}
\alias{get_implicit_value}
\title{Get implicit value (C++ style)}
\usage{
get_implicit_value(delta, charnum, iv_type = 1)

get_implicit_value(delta, charnum, iv_type = 1)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{iv_type}{Implicit value type (1 or 2, default: 1)}
}
\value{
Integer implicit value

Integer value or NA if not defined
}
\description{
Get implicit value (C++ style)

Get implicit value for a character
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
get_implicit_value(delta, 1)
}
}
