% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_item}
\alias{get_item}
\title{Get complete item data as a list}
\usage{
get_item(delta, itemnum, include_comments = FALSE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Integer item number}

\item{include_comments}{Logical; if TRUE, include comments in item names}
}
\value{
A list with item name and attributes
}
\description{
Retrieves all data for a specific item.
}
