% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_item_name}
\alias{get_item_name}
\title{Get item name (C++ style)}
\usage{
get_item_name(delta, itemnum, include_comments = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}

\item{include_comments}{Logical; if TRUE, include comments in item names}
}
\value{
Character string with item name
}
\description{
Get item name (C++ style)
}
