% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_items_debug}
\alias{get_items_debug}
\title{Get items debug output}
\usage{
get_items_debug(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with debug output
}
\description{
Get items debug output
}
