% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_items_nb}
\alias{get_items_nb}
\title{Get number of items (C++ style)}
\usage{
get_items_nb(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Integer number of items
}
\description{
Get number of items (C++ style)
}
