% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_specs_debug}
\alias{get_specs_debug}
\title{Get specifications debug output}
\usage{
get_specs_debug(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with debug output
}
\description{
Get specifications debug output
}
