% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_state}
\alias{get_state}
\title{Get state name (C++ style)}
\usage{
get_state(delta, charnum, statenum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{statenum}{State number}
}
\value{
Character string with state name
}
\description{
Get state name (C++ style)
}
