% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_states_nb}
\alias{get_states_nb}
\title{Get number of states (C++ style)}
\usage{
get_states_nb(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Integer number of states
}
\description{
Get number of states (C++ style)
}
