% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{get_version}
\alias{get_version}
\title{Get version of the underlying C++ library}
\usage{
get_version(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Version string
}
\description{
Get version of the underlying C++ library
}
