% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{has_specifications}
\alias{has_specifications}
\title{Check if specifications are available}
\usage{
has_specifications(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical
}
\description{
Check if specifications are available
}
