% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{implicit_value}
\alias{implicit_value}
\title{Get implicit value}
\usage{
implicit_value(delta, charnum, iv_type = 1)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{iv_type}{Implicit value type (1 or 2)}
}
\value{
Integer implicit value
}
\description{
Get implicit value
}
