% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{implicit_values}
\alias{implicit_values}
\title{Get all implicit values}
\usage{
implicit_values(delta, iv_type = 1)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{iv_type}{Implicit value type (1 or 2)}
}
\value{
Integer vector of implicit values
}
\description{
Get all implicit values
}
