% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{is_chars_parsed}
\alias{is_chars_parsed}
\title{Check if characters are parsed}
\usage{
is_chars_parsed(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical
}
\description{
Check if characters are parsed
}
