% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R, R/delta_methods.R
\name{is_dependent}
\alias{is_dependent}
\title{Check if character is dependent}
\usage{
is_dependent(delta, dcnum, ccnum, ccstate)

is_dependent(delta, dcnum, ccnum, ccstate)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{dcnum}{Dependent character number}

\item{ccnum}{Control character number}

\item{ccstate}{Control character state}
}
\value{
Logical indicating if dependent

Logical indicating if dcnum is dependent
}
\description{
Check if character is dependent

Check if a character is dependent
}
\examples{
\dontrun{
delta <- load_delta("chars", "items", "specs")
is_dependent(delta, 14, 13, 2)
}
}
