% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{is_items_parsed}
\alias{is_items_parsed}
\title{Check if items are parsed}
\usage{
is_items_parsed(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical
}
\description{
Check if items are parsed
}
