% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{is_specs_parsed}
\alias{is_specs_parsed}
\title{Check if specifications are parsed}
\usage{
is_specs_parsed(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical
}
\description{
Check if specifications are parsed
}
