% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{item_attributes}
\alias{item_attributes}
\title{Get item attributes as data frame}
\usage{
item_attributes(delta, itemnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}
}
\value{
Data frame with attributes
}
\description{
Get item attributes as data frame
}
