% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{item_data}
\alias{item_data}
\title{Get item data as list}
\usage{
item_data(delta, itemnum, include_comments = FALSE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}

\item{include_comments}{Logical; if TRUE, include comments in item names}
}
\value{
List with item data
}
\description{
Get item data as list
}
