% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{item_names}
\alias{item_names}
\title{Get all item names}
\usage{
item_names(delta, include_comments = FALSE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{include_comments}{Logical; if TRUE, include comments in item names}
}
\value{
Character vector of item names
}
\description{
Get all item names
}
