% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{items_df}
\alias{items_df}
\title{Get item information as a data frame}
\usage{
items_df(delta, include_comments = FALSE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{include_comments}{Logical; if TRUE, include comments in item names}
}
\value{
A data.frame with item information
}
\description{
Extracts all item information from a DELTA dataset and returns it as a
data frame.
}
