% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{items_filename}
\alias{items_filename}
\title{Get items filename}
\usage{
items_filename(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with the filename
}
\description{
Get items filename
}
