% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{load_delta}
\alias{load_delta}
\title{Load a DELTA dataset}
\usage{
load_delta(chars_file, items_file, specs_file = NULL)
}
\arguments{
\item{chars_file}{Path to the DELTA characters file}

\item{items_file}{Path to the DELTA items file}

\item{specs_file}{Optional path to the DELTA specifications file}
}
\value{
A DeltaParser Rcpp module object
}
\description{
Create a DeltaParser object from DELTA format files.
}
