% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{matches}
\alias{matches}
\title{Check if item matches values}
\usage{
matches(delta, itemnum, charnum, values, strict = TRUE, with_extrval = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{itemnum}{Item number}

\item{charnum}{Character number}

\item{values}{Numeric vector of values}

\item{strict}{Logical; strict comparison}

\item{with_extrval}{Logical; include extreme values}
}
\value{
Logical indicating if item matches
}
\description{
Check if item matches values
}
