% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{parse_attribute}
\alias{parse_attribute}
\title{Parse a DELTA attribute string}
\usage{
parse_attribute(attr_string)
}
\arguments{
\item{attr_string}{A DELTA attribute string}
}
\value{
A list with parsed components
}
\description{
Parse a DELTA attribute string
}
