% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{parse_characters}
\alias{parse_characters}
\title{Parse characters file}
\usage{
parse_characters(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical indicating success
}
\description{
Parse characters file
}
