% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{parse_delta_attribute}
\alias{parse_delta_attribute}
\title{Parse DELTA Attribute}
\usage{
parse_delta_attribute(x)
}
\arguments{
\item{x}{Character string to parse}
}
\description{
Internal function to parse DELTA attributes from the C++ module.
}
\keyword{internal}
