% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{parse_specs}
\alias{parse_specs}
\title{Parse specifications file}
\usage{
parse_specs(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Logical indicating success
}
\description{
Parse specifications file
}
