% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{Rcpp_DeltaParser-class}
\alias{Rcpp_DeltaParser-class}
\alias{Rcpp_DeltaParser}
\title{Rcpp_DeltaParser Class}
\description{
Internal class for DELTA parser objects from the C++ module.
}
\keyword{internal}
