% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_char_type}
\alias{set_char_type}
\title{Set character type}
\usage{
set_char_type(delta, charnum, chartype)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{chartype}{Integer character type code}
}
\description{
Set character type
}
