% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_char_type_by_name}
\alias{set_char_type_by_name}
\title{Set character type by name}
\usage{
set_char_type_by_name(delta, charnum, type_name)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{type_name}{Character type name}
}
\description{
Set character type by name
}
