% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_chars_filename}
\alias{set_chars_filename}
\title{Set characters filename}
\usage{
set_chars_filename(delta, fname, parse = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{fname}{New filename}

\item{parse}{Logical, parse the file immediately}
}
\description{
Set characters filename
}
