% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_filename}
\alias{set_filename}
\title{Set filename (C++ style)}
\usage{
set_filename(delta, fname, parse = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{fname}{New filename}

\item{parse}{Logical, parse the file immediately}
}
\description{
Set filename (C++ style)
}
