% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_items_file}
\alias{set_items_file}
\title{Set items filename}
\usage{
set_items_file(delta, fname, parse = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{fname}{New filename}

\item{parse}{Logical, parse the file immediately}
}
\description{
Set items filename
}
