% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{set_specs_file}
\alias{set_specs_file}
\title{Set specifications filename}
\usage{
set_specs_file(delta, fname, parse = TRUE)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{fname}{New filename}

\item{parse}{Logical, parse the file immediately}
}
\description{
Set specifications filename
}
