% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{specs_filename}
\alias{specs_filename}
\title{Get specifications filename}
\usage{
specs_filename(delta)
}
\arguments{
\item{delta}{A DeltaParser object}
}
\value{
Character string with the filename
}
\description{
Get specifications filename
}
