% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{state_count}
\alias{state_count}
\title{Get number of states for a character}
\usage{
state_count(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Integer number of states
}
\description{
Get number of states for a character
}
