% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{state_name}
\alias{state_name}
\title{Get state name}
\usage{
state_name(delta, charnum, statenum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}

\item{statenum}{State number}
}
\value{
Character string with state name
}
\description{
Get state name
}
