% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{states}
\alias{states}
\title{Get all states for a character}
\usage{
states(delta, charnum)
}
\arguments{
\item{delta}{A DeltaParser object}

\item{charnum}{Character number}
}
\value{
Character vector of state names
}
\description{
Get all states for a character
}
