% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{strip_comments}
\alias{strip_comments}
\title{Remove comments from a DELTA string}
\usage{
strip_comments(src)
}
\arguments{
\item{src}{A string with DELTA comments}
}
\value{
String with all comments removed
}
\description{
Remove comments from a DELTA string
}
