% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{strip_delta_comments}
\alias{strip_delta_comments}
\title{Strip DELTA Comments}
\usage{
strip_delta_comments(x)
}
\arguments{
\item{x}{Character string to strip comments from}
}
\description{
Internal function to strip comments from DELTA format.
}
\keyword{internal}
