% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncoded.R
\name{uncoded}
\alias{uncoded}
\title{Generate a list of uncoded characters using CONFOR}
\usage{
uncoded(delta, heading = NULL, remove_temp = TRUE)
}
\arguments{
\item{delta}{A Delta dataset object (created by \code{\link{load_delta}})}

\item{heading}{Heading for the output (default = NULL, uses heading from specs file)}

\item{remove_temp}{Logical; if TRUE, remove temporary files after conversion (default = TRUE)}
}
\value{
Invisibly returns the path to the generated uncoded characters file
}
\description{
This function creates a directives file for the CONFOR program to generate
a list of characters that are not coded (i.e., have no data) in a DELTA dataset.
This is useful for identifying missing data and gaps in taxonomic descriptions.
}
\details{
The function assumes that the CONFOR program executable is available
in the system PATH or in the current working directory.
}
\examples{
\dontrun{
# Load a DELTA dataset
delta <- load_delta("chars_viola", "items_viola", "specs_viola")

# Generate list of uncoded characters
uncoded(delta)

# With custom heading
uncoded(delta, heading = "Genre Viola - Missing data report")
}
}
