% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDelta-package.R
\docType{package}
\name{RDelta-package}
\alias{RDelta-package}
\alias{RDelta}
\title{RDelta: Parse and Query DELTA-Format Taxonomic Descriptions}
\description{
Provides an R interface to the tDelta C++ library for parsing
DELTA (DEscription Language for TAxonomy) format files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maurobio/DeltaR}
  \item Report bugs at \url{https://github.com/maurobio/DeltaR/issues}
}

}
\author{
\strong{Maintainer}: Mauro Cavalcanti \email{maurobio@gmail.com} (Original tDelta C++ library by Denis Ziegler)

Authors:
\itemize{
  \item Mauro Cavalcanti \email{maurobio@gmail.com} (Original tDelta C++ library by Denis Ziegler)
}

Other contributors:
\itemize{
  \item Edinaldo Nelson dos Santos Silva \email{nelson@inpa.gov.br} [contributor]
}

}
\keyword{internal}
